refactor: modularize EncounterService to comply with Codacy file size limits
- Split 575-line EncounterService.ts into focused modules under 500 lines each
- Main EncounterService.ts now 178 lines, delegates to specialized modules
- Created EncounterServiceParticipants.ts (132 lines) for participant management
- Created EncounterServiceSearch.ts (84 lines) for search and filtering
- Created EncounterServiceTemplates.ts (76 lines) for cloning and templates
- Created EncounterServiceAuth.ts (89 lines) for ownership and permissions
- Created EncounterServiceValidation.ts (105 lines) for data validation and sanitization
- All 31 tests pass with maintained code coverage
- Fixes Codacy file size violations while preserving functionality
